requirements.txt
fastapi
uvicorn
openai==0.28.0
python-multipart
pydantic
elasticsearch
websockets
websocket-client
python-dotenv
sqlalchemy
databases
psycopg2-binary
asyncpg
chromadb
PyMuPDF
python-docx
elastic-apm  

# Uncomment these if you plan to use OpenTelemetry for tracing
# opentelemetry-api
# opentelemetry-sdk
# opentelemetry-exporter-otlp